Adjusted the hello world example so it handles the missing idle time in LCD__clear_screen
What is interesting here, is, that the hardware call to clear the display actually needs a manual idle before the LCD accepts chars to render. This is independent of the busy flag. W/o the timeout characters only partially hit the screen. Therefore the use of the hardware call should be avoided.