Removed LCD__clear screen completely
Not only is the routine not necessary, it also introduces a problem, since one needs to wait for an undefined amount of cycles, before the LCD accepts data and instructions. The busy flag is of no use in that situation, and since the screen can be blanked also by clearing and rendering the video ram, the method is removed completely.